requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
pycodestyle==2.9.1
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
